+++ crates/gpui/examples/stickies/main.rs
@@ -1,10 +1,13 @@
 use gpui::{
-    AnyWindowHandle, App, Application, Bounds, Context, CursorStyle, ElementId, Entity,
-    FocusHandle, Focusable, Hsla, KeyBinding, Menu, MenuItem, Pixels, Point, SharedString,
-    SystemMenuType, Window, WindowBounds, WindowKind, WindowOptions, actions, div, prelude::*, px,
-    rgb, size,
+    App, AppContext, Application, Bounds, KeyBinding, Menu, MenuItem, Pixels, Point, SharedString,
+    SystemMenuType, WindowBounds, WindowKind, WindowOptions, actions, px, size,
 };
 use std::sync::atomic::{AtomicUsize, Ordering};
+
+mod text_area;
+mod ui;
+
+use ui::{Sticky, StickyColor};
 
 actions!(
     stickies,
@@ -60,18 +63,18 @@ actions!(
 static STICKY_COUNT: AtomicUsize = AtomicUsize::new(0);
 
 // Window setup
-// - [ ] open two stickies, one blue, one yellow
-// - [ ] blue behind, yellow in front
-// - [ ] black border around window
+// - [x] open two stickies, one blue, one yellow
+// - [x] blue behind, yellow in front
+// - [x] black border around window
 
 // Titlebar
-// - [ ] only focused window shows titlebar
-// - [ ] titlebar is darker version of content area color
+// - [x] only focused window shows titlebar
+// - [x] titlebar is darker version of content area color
 // - [ ] double-clicking the titlebar minimizes it (and shows a tiny snippet of the content)
 // - [ ] (left) - square closes app, (right) - triangle zooms, - minus icon collapses/expands sticky
 
 // Content area
-// - [ ] content area is themed by a color, one of 6 options: blue, green, yellow, pink, purple, gray
+// - [x] content area is themed by a color, one of 6 options: blue, green, yellow, pink, purple, gray
 
 // Menubar
 // - [ ] has a menubar with:
@@ -111,272 +114,17 @@ Look in Help to find out more about using Stickies.
 
 struct TextArea {}
 
-#[derive(Clone, Default, Debug)]
-enum StickyColor {
-    #[default]
-    Yellow,
-    Blue,
-    Green,
-    Pink,
-    Purple,
-    Gray,
-}
-
-impl StickyColor {
-    fn bg(&self) -> Hsla {
-        match self {
-            StickyColor::Yellow => rgb(0xFFF48F).into(),
-            StickyColor::Blue => rgb(0x98F6FF).into(),
-            StickyColor::Green => rgb(0x9BFF88).into(),
-            StickyColor::Pink => rgb(0xFFB3E0).into(),
-            StickyColor::Purple => rgb(0xD0B3FF).into(),
-            StickyColor::Gray => rgb(0xD4D4D4).into(),
-        }
-    }
-
-    fn titlebar_bg(&self) -> Hsla {
-        match self {
-            StickyColor::Yellow => rgb(0xFFE900).into(),
-            StickyColor::Blue => rgb(0x5DF3FF).into(),
-            StickyColor::Green => rgb(0x6FFF52).into(),
-            StickyColor::Pink => rgb(0xFF8FD0).into(),
-            StickyColor::Purple => rgb(0xB88FFF).into(),
-            StickyColor::Gray => rgb(0xB0B0B0).into(),
-        }
-    }
-
-    /// the cmd+number modifier for this color
-    /// used to bind the color changing shortcuts
-    fn cmd_number(&self) -> u8 {
-        match self {
-            StickyColor::Yellow => 1,
-            StickyColor::Blue => 2,
-            StickyColor::Green => 3,
-            StickyColor::Pink => 4,
-            StickyColor::Purple => 5,
-            StickyColor::Gray => 6,
-        }
-    }
-}
-
-struct Sticky {
-    id: ElementId,
-    focus_handle: FocusHandle,
-    bounds: Bounds<Pixels>,
-    color: StickyColor,
-    collapsed: bool,
-
-    content: SharedString,
-    window_handle: Option<AnyWindowHandle>,
-    // text_area: Entity<TextArea>,
-}
-
-impl Sticky {
-    pub fn new(
-        cx: &mut App,
-        id: impl Into<ElementId>,
-        bounds: Bounds<Pixels>,
-        color: StickyColor,
-    ) -> Self {
-        Self {
-            id: id.into(),
-            focus_handle: cx.focus_handle(),
-            bounds,
-            color,
-            collapsed: false,
-            content: SharedString::new(""),
-            window_handle: None,
-        }
-    }
-
-    pub fn content(mut self, content: impl Into<SharedString>) -> Self {
-        self.content = content.into();
-        self
-    }
-
-    pub fn with_window_handle(mut self, handle: AnyWindowHandle) -> Self {
-        self.window_handle = Some(handle);
-        self
-    }
-
-    fn close_note(&mut self, _: &CloseNote, window: &mut Window, _cx: &mut Context<Self>) {
-        window.remove_window();
-    }
-
-    fn zoom(&mut self, _: &ZoomWindow, window: &mut Window, _cx: &mut Context<Self>) {
-        window.zoom_window();
-    }
-
-    fn change_color(&mut self, color: StickyColor, _window: &mut Window, cx: &mut Context<Self>) {
-        self.color = color;
-        cx.notify();
-    }
-
-    fn new_note(&mut self, _: &NewNote, window: &mut Window, cx: &mut Context<Self>) {
-        let current_bounds = window.window_bounds().get_bounds();
-        let screen_bounds = window.display(cx).map(|d| d.bounds()).unwrap_or(Bounds {
-            origin: Point::default(),
-            size: size(px(1920.), px(1080.)),
-        });
-
-        let offset = px(24.);
-        let new_size = size(px(DEFAULT_STICKY_SIZE.0), px(DEFAULT_STICKY_SIZE.1));
-
-        // Try to place new note below current one
-        let mut new_origin = Point {
-            x: current_bounds.origin.x + offset,
-            y: current_bounds.origin.y + current_bounds.size.height + offset,
-        };
-
-        // If it would go off the bottom of the screen, place it above instead
-        if new_origin.y + new_size.height > screen_bounds.bottom() {
-            new_origin.y = current_bounds.origin.y - new_size.height - offset;
-        }
-
-        let new_bounds = Bounds {
-            origin: new_origin,
-            size: new_size,
-        };
-
-        let sticky_id = STICKY_COUNT.fetch_add(1, Ordering::SeqCst) as u64;
-
-        cx.open_window(
-            WindowOptions {
-                window_bounds: Some(WindowBounds::Windowed(new_bounds)),
-                titlebar: None,
-                window_background: gpui::WindowBackgroundAppearance::Transparent,
-                focus: true,
-                show: true,
-                kind: WindowKind::Normal,
-                ..Default::default()
-            },
-            |window, cx| {
-                cx.new(|cx| {
-                    let window_handle = window.window_handle();
-                    Sticky::new(
-                        cx,
-                        ElementId::NamedInteger("sticky".into(), sticky_id),
-                        new_bounds,
-                        StickyColor::Yellow,
-                    )
-                    .with_window_handle(window_handle.into())
-                })
-            },
-        )
-        .unwrap();
-    }
-}
-
-impl Render for Sticky {
-    fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        let entity = cx.entity();
-        let window_active = window.is_window_active();
-        let focus_handle = self.focus_handle.clone();
-
-        div()
-            .id(self.id.clone())
-            .key_context("Note")
-            .track_focus(&self.focus_handle(cx))
-            .cursor(CursorStyle::IBeam)
-            .on_action(cx.listener(Self::close_note))
-            .on_action(cx.listener(Self::new_note))
-            .on_action(cx.listener(Self::zoom))
-            .on_action(cx.listener(|this, _: &ChangeColorYellow, window, cx| {
-                this.change_color(StickyColor::Yellow, window, cx)
-            }))
-            .on_action(cx.listener(|this, _: &ChangeColorBlue, window, cx| {
-                this.change_color(StickyColor::Blue, window, cx)
-            }))
-            .on_action(cx.listener(|this, _: &ChangeColorGreen, window, cx| {
-                this.change_color(StickyColor::Green, window, cx)
-            }))
-            .on_action(cx.listener(|this, _: &ChangeColorPink, window, cx| {
-                this.change_color(StickyColor::Pink, window, cx)
-            }))
-            .on_action(cx.listener(|this, _: &ChangeColorPurple, window, cx| {
-                this.change_color(StickyColor::Purple, window, cx)
-            }))
-            .on_action(cx.listener(|this, _: &ChangeColorGray, window, cx| {
-                this.change_color(StickyColor::Gray, window, cx)
-            }))
-            .relative()
-            .bg(self.color.bg())
-            .border_1()
-            .border_color(gpui::black())
-            .w(self.bounds.size.width)
-            .h(self.bounds.size.height)
-            .flex_col()
-            .text_size(px(12.))
-            .line_height(px(14.))
-            .pt(px(TITLEBAR_HEIGHT)) // reserve space for absolutely positioned titlebar
-            .on_click(cx.listener(move |_, _, window, cx| {
-                if !window.is_window_active() {
-                    window.activate_window();
-                }
-                window.focus(&focus_handle);
-                cx.notify();
-            }))
-            .child(Titlebar::new(entity, window_active))
-            .child(
-                div()
-                    .flex_1()
-                    .py(px(8.))
-                    .px(px(14.))
-                    .child(self.content.clone()),
-            )
-    }
-}
-
-impl Focusable for Sticky {
-    fn focus_handle(&self, _cx: &App) -> gpui::FocusHandle {
-        self.focus_handle.clone()
-    }
-}
-
-#[derive(IntoElement)]
-struct Titlebar {
-    sticky: Entity<Sticky>,
-    window_active: bool,
-}
-
-impl Titlebar {
-    pub fn new(sticky: Entity<Sticky>, window_active: bool) -> Self {
-        Self {
-            sticky,
-            window_active,
-        }
-    }
-}
-
-impl RenderOnce for Titlebar {
-    fn render(self, _window: &mut Window, cx: &mut App) -> impl IntoElement {
-        let sticky_state = self.sticky.read(cx);
-        let color = sticky_state.color.clone();
-
-        div()
-            .absolute()
-            .top_0()
-            .left_0()
-            .h(px(12.))
-            // todo: probably needs to get the width from `sticky_state`
-            .w_full()
-            .when(self.window_active, |this| this.bg(color.titlebar_bg()))
-    }
-}
-
 fn setup_menus(cx: &mut App) {
     cx.set_menus(vec![
         Menu {
             name: "File".into(),
             items: vec![
                 MenuItem::action("New Note", NewNote),
+                MenuItem::separator(),
                 MenuItem::action("Close", CloseNote),
                 MenuItem::separator(),
                 MenuItem::action("Import Text...", ImportText),
                 MenuItem::action("Export Text...", ExportText),
-                MenuItem::separator(),
-                MenuItem::action("Page Setup...", PageSetup),
-                MenuItem::action("Print...", Print),
             ],
         },
         Menu {
